root: re-brand it to my personal project completely
Since there is a need to shift from GitLab to Github due to their
pricing changes and ZORALab does not want to maintain this project
in the future, I should shift and re-brand it into my personal
project completely under my GitHub profile. Hence, let's do it.

This patch re-brands the repository into my personal project
completely and shall be updated somewhere in the near future.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>